cnn_lstm_kmeans_anomaly_2-3_axis_dataset/sierra_modified/requirement.txt
absl-py                       0.9.0      
astor                         0.8.1      
astunparse                    1.6.3      
bokeh                         1.4.0      
cachetools                    4.0.0      
certifi                       2019.11.28 
chardet                       3.0.4      
cycler                        0.10.0     
decorator                     4.4.2      
dm-tree                       0.1.5      
future                        0.18.2     
gast                          0.2.2      
google-auth                   1.11.2     
google-auth-oauthlib          0.4.1      
google-pasta                  0.1.8      
grpcio                        1.26.0     
h5py                          2.10.0     
idna                          2.9        
Jinja2                        2.11.1     
joblib                        0.14.1     
Keras                         2.3.1      
Keras-Applications            1.0.8      
Keras-Preprocessing           1.1.0      
keras-tcn                     3.0.1      
kiwisolver                    1.2.0      
larq                          0.9.6      
larq-compute-engine           0.1.2      
Markdown                      3.1.1      
MarkupSafe                    1.1.1      
matplotlib                    3.2.1      
networkx                      2.4        
numpy                         1.18.1     
oauthlib                      3.1.0      
opt-einsum                    3.1.0      
packaging                     20.1       
pandas                        0.24.2     
parse                         1.6.5      
Pillow                        6.0.0      
pip                           19.2.3     
protobuf                      3.11.2     
pyasn1                        0.4.8      
pyasn1-modules                0.2.8      
pyparser                      1.0        
pyparsing                     2.4.6      
python-dateutil               2.8.1      
pytz                          2019.3     
PyYAML                        5.3        
QKeras                        0.8.0      
requests                      2.23.0     
requests-oauthlib             1.3.0      
rsa                           4.0        
scikit-learn                  0.20.3     
scipy                         1.4.1      
seaborn                       0.10.1     
setuptools                    41.2.0     
six                           1.14.0     
sklearn                       0.0        
tensorboard                   2.0.2      
tensorboard-plugin-wit        1.6.0.post3
tensorflow-estimator          2.0.1      
tensorflow-gpu                2.0.0      
tensorflow-model-optimization 0.3.0      
termcolor                     1.1.0      
terminaltables                3.1.0      
torch                         1.5.0      
tornado                       6.0.3      
tqdm                          4.43.0     
urllib3                       1.25.8     
Werkzeug                      0.16.1     
wheel                         0.34.1     
wrapt                         1.11.2  